Moved dashboard to jobs_controller.
