+++ app/views/jobs/index.html.erb
@@ -3,7 +3,8 @@
 <% end %>
 
 <p>
-This page shows you the status of current and recent autograding jobs. 
+This page shows you the status of current and recent autograding jobs.<br />
+<%= link_to "Click here", tango_status_course_jobs_path %> to view the overall status of Tango VMs.
 </p>
 
 <h2>Currently Running Jobs</h2>
